requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.8.2
python-dotenv==1.0.1
httpx==0.27.2
python-multipart==0.0.9
numpy==2.1.1
pypdf==4.3.1
beautifulsoup4==4.12.3
google-generativeai==0.7.2
